Reimplement vdeplug_python using ctypes.
ctypes allows to write extensions in pure Python. It is also part of the
Python's standard library and more portable across different
implementations and versions. This is also a first step to support
Python3.

A compatibility layer to the old vdeplug_python is also supplied as like
to the VdePlug module.